+++ R/model_info.R
@@ -28,7 +28,7 @@
 #'      \item \code{is_tweedie}: family is tweedie
 #'      \item \code{is_ordinal}: family is ordinal, multinomial, or cumulative link
 #'      \item \code{is_cumulative}: family is ordinal, multinomial, or cumulative link
-#'      \item \code{is_multinomial}: family is ordinal, multinomial, categorical, or cumulative link
+#'      \item \code{is_multinomial}: family is multinomial or categorical link
 #'      \item \code{is_categorical}: family is categorical link
 #'      \item \code{is_censored}: model is a censored model (has a censored response, including survival models)
 #'      \item \code{is_truncated}: model is a truncated model (has a truncated response)
@@ -654,13 +654,16 @@ model_info.LORgee <- function(x, ...) {
     link <- "logit"
   }
 
-  faminfo <- stats::binomial(link = link)
+  if (x$link == "Cumulative logit")
+    family <- "ordinal"
+  else
+    family <- "multinomial"
 
   .make_family(
     x = x,
-    fitfam = faminfo$family,
-    logit.link = faminfo$link == "logit",
-    link.fun = faminfo$link,
+    fitfam = family,
+    logit.link = link == "logit",
+    link.fun = link,
     ...
   )
 }